+++ src/utils/utils.js
@@ -1,33 +1,35 @@
+import chalk from "chalk";
 import figlet from "figlet";
 import gradient from "gradient-string";
 import pressAnyKey from "press-any-key";
 import shell from "shelljs";
 import menu from "../cli.js";
-import chalk from "chalk";
 
 export const colors = {
   accent: chalk.keyword("lightblue").bold,
-  errorColor: chalk.keyword("tomato").bold.underline,
-  title: gradient("yellow", "tomato", "cyan"),
-  keyword: "aquamarine",
-  menuLine: chalk.redBright,
-  spacer: chalk.blueBright,
-  bold: chalk.bold,
-  step: chalk.keyword("aquamarine").bold,
-  frames: chalk.keyword("tomato").bold("Frames"),
   author: chalk.keyword("darkorange").italic.bold("Guillermo Brachetta."),
-  url: chalk.cyanBright.bold("https://github.com/GBrachetta/frames"),
-  pipenv: chalk.cyan.bold("Pipenv"),
-  vsc: chalk.cyan.bold("Visual Studio Code"),
-  help: chalk.keyword("powderblue"),
-  ok: chalk.greenBright.bold,
+  bold: chalk.bold,
+  errorColor: chalk.keyword("tomato").bold.underline,
+  exit: chalk.keyword("tomato").bold,
   frame: {
+    Django: chalk.keyword("yellowgreen").bold,
+    "Next.js": chalk.keyword("springgreen").bold,
     React: chalk.keyword("cyan").bold,
     "Vite-React": chalk.keyword("gold").bold,
     "Vite-Vue": chalk.keyword("hotpink").bold,
-    "Next.js": chalk.keyword("springgreen").bold,
-    Django: chalk.keyword("yellowgreen").bold,
   },
+  frames: chalk.keyword("tomato").bold("Frames"),
+  help: chalk.keyword("powderblue"),
+  helpMenu: chalk.keyword("springgreen").bold,
+  keyword: "aquamarine",
+  menuLine: chalk.redBright,
+  ok: chalk.greenBright.bold,
+  pipenv: chalk.cyan.bold("Pipenv"),
+  spacer: chalk.blueBright,
+  step: chalk.keyword("aquamarine").bold,
+  title: gradient("yellow", "tomato", "cyan"),
+  url: chalk.cyanBright.bold("https://github.com/GBrachetta/frames"),
+  vsc: chalk.cyan.bold("Visual Studio Code"),
 };
 
 const { frames, author, url, pipenv, vsc, help, ok, accent, errorColor } =