+++ src/utils/utils.js
@@ -2,13 +2,11 @@ import chalk from "chalk";
 import figlet from "figlet";
 import gradient from "gradient-string";
 import pressAnyKey from "press-any-key";
-import shell from "shelljs";
 import menu from "../cli.js";
 
 export const colors = {
   accent: chalk.keyword("lightblue").bold,
   author: chalk.keyword("darkorange").italic.bold("Guillermo Brachetta."),
-  bold: chalk.bold,
   errorColor: chalk.keyword("tomato").bold.underline,
   exit: chalk.keyword("tomato").bold,
   frame: {
@@ -21,65 +19,85 @@ export const colors = {
   frames: chalk.keyword("tomato").bold("Frames"),
   help: chalk.keyword("powderblue"),
   helpMenu: chalk.keyword("springgreen").bold,
+  helpAccent: chalk.keyword("navajowhite").bold,
   keyword: "aquamarine",
-  menuLine: chalk.redBright,
-  ok: chalk.greenBright.bold,
-  pipenv: chalk.cyan.bold("Pipenv"),
-  spacer: chalk.blueBright,
+  menuLine: chalk.keyword("crimson"),
+  spacer: chalk.keyword("lightseagreen"),
   step: chalk.keyword("aquamarine").bold,
   title: gradient("yellow", "tomato", "cyan"),
-  url: chalk.cyanBright.bold("https://github.com/GBrachetta/frames"),
-  vsc: chalk.cyan.bold("Visual Studio Code"),
 };
 
-const { frames, author, url, pipenv, vsc, help, ok, accent, errorColor } =
-  colors;
+const {
+  frames,
+  author,
+  help,
+  helpMenu,
+  accent,
+  errorColor,
+  helpAccent,
+  title,
+} = colors;
 
 export const helpMe = () => {
-  shell.echo(help(`\n${frames} scaffolds apps in different flavors.`));
-  shell.echo(
-    help(`\nIf ${vsc} is installed, it starts it once the all is done.`)
+  console.log(help(`\n  ${frames} scaffolds apps in different flavors.`));
+  console.log(
+    help(
+      `\n  If ${helpAccent(
+        "VSCode"
+      )} is installed, it starts it once the all is done.`
+    )
   );
-  shell.echo(
-    help(`\n${frames} checks for ${pipenv} and eventually installs it.`)
+  console.log(
+    help(
+      `\n  ${frames} checks for ${helpAccent(
+        "Pipenv"
+      )} and eventually installs it.`
+    )
   );
-  shell.echo(help(`\nRun ${frames} in the parent directory of your app.`));
-  shell.echo(help(`\n${frames} was created by ${author}`));
-  shell.echo(help(`\nFeel free to use, share and contribute: ${url}\n`));
-
-  pressAnyKey(ok("Press any key to go back to the menu...")).then(() => menu());
+  console.log(help(`\n  Run ${frames} in the parent directory of your app.`));
+  console.log(help(`\n  ${frames} was created by ${author}`));
+  console.log(
+    help(
+      `\n  Feel free to use, share and contribute: ${helpAccent(
+        "https://github.com/GBrachetta/frames"
+      )}\n`
+    )
+  );
+  pressAnyKey(helpMenu("  Press any key to go back to the menu...")).then(() =>
+    menu()
+  );
 };
 
 export const goodbye = (framework, framecolor, name) => {
-  const bye = gradient("yellow", "tomato", "cyan");
-
   if (!name) {
-    shell.echo();
-    shell.echo(bye(figlet.textSync("  Bye!")));
-    shell.echo();
+    console.log();
+    console.log(title(figlet.textSync("  Bye!")));
+    console.log();
     return;
   }
   const path = framecolor(`${process.cwd()}/${name}`);
   const project = `${framecolor(framework)} project`;
-  shell.echo(ok("\nAll done!\n"));
-  shell.echo(
+  console.log(helpMenu("\nAll done!\n"));
+  console.log(
     accent(
       `Your ${project} is ready at ${path}. Now go and create a great app!`
     )
   );
-  shell.echo();
-  shell.echo(bye(figlet.textSync("  Bye!")));
-  shell.echo();
+  console.log();
+  console.log(title(figlet.textSync("  Bye!")));
+  console.log();
 };
 
 export const noPython = () => {
-  shell.echo(errorColor("\nMust have Python > 3.6 installed to run Django!\n"));
+  console.log(
+    errorColor("\nMust have Python > 3.6 installed to run Django!\n")
+  );
   process.exit(1);
 };
 
 export const failedPipenv = (error) => {
-  shell.echo(errorColor("Failed to install Pipenv, please install!"));
-  shell.echo(ok("\nVisit: https://pypi.org/project/pipenv/"));
-  shell.echo(errorColor("\nMessage:"), accent(error.message), "\n");
+  console.log(errorColor("Failed to install Pipenv, please install!"));
+  console.log(helpMenu("\nVisit: https://pypi.org/project/pipenv/"));
+  console.log(errorColor("\nMessage:"), accent(error.message), "\n");
   process.exit(1);
 };