+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.2.2] - 2021-05-28
+
+- Improve Regex to check for valid directory name
+- Further optimize centralization of color variables
+
 ## [3.2.1] - 2021-05-28
 
 - Fix `.vscode` folder not being copied to Django template